feat/approval: #109 결재 문서 생성 기능 구현 중 ....

+++ perflow-backend/src/main/java/com/touchdown/perflowbackend/approval/query/dto/ApproveLineDetailDTO.java
@@ -14,7 +14,7 @@ public class ApproveLineDetailDTO {
 
     private final ApproveType approveType;
 
-    private final Integer approveLineOrder;
+    private final Long approveLineOrder;
 
     private final List<ApproveSbjDTO> approveSbjs;
 }
